[builder][FIX] Detect library namespace automatically (https://github.com/SAP/ui5-builder/pull/255)
Like already done for applications, read namespace information from "sap.app".id
attribute of manifest.json.

Before, some tasks derived the namespace from a libraries name.
This contract was not documented and not enforced. Causing some confusion when
using those tasks.

If no manifest.json is present, fall back to library.name attribute of .library.
If .library is not present, fall back to the relative path of the library.js file.
If that is not present, the namespace is undefined and some tasks are never
executed for that library:
- generateJsdoc
- executeJsdocSdkTransformation
- generateManifestBundle

namespaces might become mandatory in a future major release.