[cli][FIX] serve: Create SSL certificate in user homedir
The default certificate key/cert paths were using the $HOME variable
which was not replaced with the actual homedir of the user.
This caused creating a "$HOME" folder within the project which was not
intended.
Furthermore this can be dangerous in case someone wants to remove that
folder with `rm -rf`, as this might cause the homedir to be removed in
case no single quotes are used.