[cli][FEATURE] CLI: Improve error reporting
Yargs errors (like 'Unkown argument') do not print the command help
anymore but a hint to use 'ui5 --help'.

Exceptions are now formatted in a nice way.

On default log level only the error message is printed out and a hint to
use the --verbose flag for details.

On verbose log level, the stack trace is printed as well and a hint
featuring the URL to open a new issue for the UI5 Tooling module that
is most likely at fault based on an analysis of the first line of the
stack trace.