[builder][INTERNAL] lbt escapePropertiesFile: Expect *.properties files in UTF-8 by default
If the project a "*.properties" resource originates from cannot be
determined, or if the project does not define a
propertiesFileSourceEncoding configuration or uses a legacy specVersion
(<2.0), escapePropertiesFile assumes that the resource is UTF-8
encoded instead of ISO-8859-1.

This is an internal module that *should* be wrapped by APIs that enforce
proper projects which always define a propertiesFileSourceEncoding
configuration. Therefore this change is assumed to not be breaking.