[builder][FIX] Bundler: Sort raw modules by default
Unlike the documentation states, topological sorting of modules within
'raw' sections was not enabled by default.

Sorting is based on dependencies only, so enabling it by default should
only solve issues in case dependencies are declared but sort was not
enabled.
If no dependencies are defined, the order will not change.

This could only break use cases where dependencies have been wrongly
declared and a user relies on the current unsorted order.