[fs][INTERNAL] Resource: Add getName function
Modules like the "serveIndex" middleware or the "fsInterface" currently
determine the name of a Resource by using the private _name attribute.
This is typically required for listing the content of a directory.

This attribute was missing from ResourceFacade, which lead to issues in
both modules.

In addition to adding the missing, private attribute. Also add a public
function getName().