[INTERNAL] Create versioning for the documentation (#697)
JIRA: CPOUI5FOUNDATION-565

This change creates a new documentation release workflow that enables versioning.

It extends the MkDocs squidfunk/mkdocs-material image by adding mike to it.
The GitHub Actions workflow is also being extended to support the new feature.

URLs of the documentation would be changed with a version number prepending the path. In order to "keep" old links available for the end users, there's a custom redirect for GH Pages.