[INTERNAL] Fix infinite looping for unknown sources (#747)
This is an issue with the Docs. With the implementation of versioned documentation, we wanted to support also the legacy links. That's why, when a resource is not found we check whether this resource is in the versioned documentation and if not, we redirect to the same path of the stable release. If the resource is already versioned, but is not found, we just redirect to the stable release's root path.