[builder][FIX] generateJsdoc: Also process resources created by preceeding tasks
End the use of byGlobSource and alwasy use byGlob. None of the standard
tasks executed before it add any resources. So no overhead is to be
expected from this.

On the other hand, the use of custom task extensions (for exmaple for
transpilation) requires the generateJsdoc to process resources the task
has created.

This was the last usage of the byGlobSource API of the DuplexCollection.
Therefore the API shall be deprecated in ui5-fs and removed in the
future.