[project][FIX] ProjectBuilder now can be executed in parallel  (https://github.com/SAP/ui5-project/pull/669)
Fixes: https://github.com/SAP/ui5-tooling/issues/887

ProjectBuilder was using a singleton of the logger module. Basically,
the logger was constructed only once, during module evaluation, but used
throughout the whole module.
The tricky part is that the `ProjectBuilder`'s logger has this
`setProjects()` method that acts as an allow list of the projects being
built. As the logger used to be a singleton within this module/class and
the async behaviour of the whole module, calling `setProject()` was
polluting the internal `projects` property.

With this fix, every `ProjectBuilder` instance has its own`BuildLogger`
instance that solves the issue.