[builder][INTERNAL] lib/processors/jsdoc: remove duplicates after `processingComplete`
Processing of @augments, @mixin and @borrows tags might produce new
duplicates which are not yet visible during the `parseComplete` event.

Moving the cleanup to `processingComplete` fixes this.

The merge of the additional data that the UI5 AST visitor collected,
is still done in `parseComplete` so that it is available when doclets
are cloned during the processing of @augments, @mixin and @borrows tags.

Cherry-picked from SAP/openui5@96f6f095d.