[builder][FIX] manifestEnhancer: Only use valid files for supportedLocales (https://github.com/SAP/ui5-builder/pull/1080)
This fixes two problems that could have occurred:

A properties file with an invalid locale was still taken into the list
of supported locales, which then caused a runtime exception in the
ResourceBundle as it validates the input.

Another problem was that properties files could have a valid name
according to BCP47, but the file won't be ever requested with that name.
This is due to the fact that the ResourceBundle does use the legacy Java
locale format (using underscores instead of dashes) for the request URL.

In both cases, the properties file is now ignored and no entry for the
supportedLocales is created.

Only locales that are valid according to the legacy Java locale format
are considered.
However, there is one special case: sr_Latn is also requested by the UI5
runtime, although it contains a BCP47 script, which is not valid
according to the legacy Java locale format.

---------

Co-authored-by: Merlin Beutlberger <m.beutlberger@sap.com>