[builder][INTERNAL] lib/processors/jsdoc: take type parameters into account when validating api.json
When validating classes or methods, the JSDoc template so far complained
about usages of type parameters as they could not be found in the set
of known types.

With this change, type parameters are made known in the scope in which
they are defined (method or class).

Cherry-picked from UI5/openui5@073050a9b.